Minor fix
* Fix lỗi tính toán số bước Interpolation Search
* Fix lỗi không hiện đồ thị hashing
* Add Filter tốt hơn cho Hashing array

+++ AVS/src/java/GraphController/SearchStepServlet.java
@@ -40,8 +40,10 @@ public class SearchStepServlet extends HttpServlet {
 
             // Probing the position with keeping 
             // uniform distribution in mind. 
-            int pos = low + (((high - low)
-                    / (arr[high] - arr[low])) * (searchvalue - arr[low]));
+//            int pos = (int) (low + ((float)((float)(high - low)
+//                    / (float)(arr[high] - arr[low])) * (searchvalue - arr[low])));
+            int pos = (int) (low + (((double)(high - low) / 
+                    (arr[high] - arr[low])) * (searchvalue - arr[low]))); 
 
             count++;
             // Condition of target found
